Optimized code, and fixed a rather interesting issue:
Confusing, but basically, the way it was written before, if you leveled up past an action's level, but hadn't done the job quest for it yet, the move got upgraded anyway and was therefore unusable (if it is a move upgrade trait, not a further part of the combo). I just fixed that, so anyone leveling MCH/PLD/SMN/WAR/etc will have an easier time.